[ESSI-688] Modularizes structural behavior for controller, presenter
