+++ app/controllers/hyrax/paged_resources_controller.rb
@@ -8,6 +8,7 @@ module Hyrax
     include ESSI::WorksControllerBehavior
     include ESSI::PagedResourcesControllerBehavior
     include ESSI::RemoteMetadataLookupBehavior
+    include ESSI::StructureBehavior
     include Hyrax::BreadcrumbsForWorks
     include ESSI::BreadcrumbsForWorks
     self.curation_concern_type = ::PagedResource
@@ -31,26 +32,10 @@ module Hyrax
       params[:highlight] = search_term&.flatten&.first
     end
 
-    def structure
-      parent_presenter
-      @members = presenter.member_presenters
-      @logical_order = presenter.logical_order_object
-    end
-
     def additional_response_formats(wants)
       wants.uv do
         presenter && parent_presenter
         render 'viewer_only.html.erb', layout: 'boilerplate', content_type: 'text/html'
-      end
-    end
-
-    def save_structure
-      structure = { "label": params["label"], "nodes": params["nodes"] }
-      if curation_concern.lock?
-        head 423
-      else
-        SaveStructureJob.perform_later(curation_concern, structure.to_json)
-        head 200
       end
     end
   end
